requirements.txt
asgiref==3.5.0
Django==4.0.2
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
jdatetime==4.0.0
Pillow==9.0.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
